refactor: rename hooks folder
